requirements.txt
# === Backend API ===
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
bcrypt==4.0.1

# === Database ===
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# === Data & ML ===
pandas>=2.1.0
numpy>=1.24.0
scikit-learn==1.6.1
xgboost>=2.0.0

# === Scraping ===
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
tqdm>=4.66.0
tenacity>=8.2.0

# === Environment ===
python-dotenv>=1.0.0